Fix result collecting race condition
If multiple cells were updated in rapid succession, the result of one of
the snapshots would override the other, causing incorrect results. This
change forces the results table to draw directly from the player table,
which fixes the race condition.

Also put each user display name into the game at the start, so that
collecting the results can be done quicker as the display name doesn't
need to be fetched each time.

Signed-off-by: Matthew Gill <matthew.gill@live.co.uk>